Make disklabel(8) aware of the new fields, allowing for larger than 2T disks, partitions and offsets. ok deraadt@

+++ sbin/disklabel/extern.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: extern.h,v 1.5 2007/03/02 02:29:13 krw Exp $	*/
+/*	$OpenBSD: extern.h,v 1.6 2007/06/08 18:42:17 otto Exp $	*/
 
 /*
  * Copyright (c) 2003 Theo de Raadt <deraadt@openbsd.org>
@@ -18,8 +18,8 @@
 
 u_short	dkcksum(struct disklabel *);
 int	checklabel(struct disklabel *);
-double	scale(u_int32_t, char, struct disklabel *);
-void	display(FILE *, struct disklabel *, char **, char, int, u_int32_t);
+double	scale(u_int64_t, char, struct disklabel *);
+void	display(FILE *, struct disklabel *, char **, char, int, u_int64_t);
 void	display_partition(FILE *, struct disklabel *, char **, int, char);
 
 struct disklabel *readlabel(int);
